Milestone/Phase2/P1/streamlit_deploy/requirements.txt
streamlit
sklearn
pandas
joblib
numpy
tensorflow-cpu